move prng methods to .cpp file
